+++ admin/includes/view_all_comments.php
@@ -74,36 +74,39 @@
  </table>
 
 
-<?php // DELETING COMMENTS
+<?php 
 
-if (isset($_GET['delete'])) {
-    $the_comment_id = escape($_GET['delete']);
+if ($_SESSION['user_role'] == 'admin') {
+  
+    // DELETING COMMENTS
+    if (isset($_GET['delete'])) {
+        $the_comment_id = escape($_GET['delete']);
 
-    $query = "DELETE FROM comments WHERE comment_id = {$the_comment_id}";
+        $query = "DELETE FROM comments WHERE comment_id = {$the_comment_id}";
 
-    $delete_query = mysqli_query($connection, $query);
-    header("Location: comments.php");
+        $delete_query = mysqli_query($connection, $query);
+        header("Location: comments.php");
+    }
+
+    // UPDATING APPROVAL STATUS to UNAPPROVED
+    if (isset($_GET['unapprove'])) {
+        $the_comment_id = escape($_GET['unapprove']);
+
+        $query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = {$the_comment_id} ";
+
+        $unapprove_comment_query = mysqli_query($connection, $query);
+        header("Location: comments.php");
+    }
+
+    // UPDATING APPROVAL STATUS to APPROVED
+    if (isset($_GET['approve'])) {
+        $the_comment_id = escape($_GET['approve']);
+
+        $query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = {$the_comment_id} ";
+
+        $approve_comment_query = mysqli_query($connection, $query);
+        header("Location: comments.php");
+    }
 }
-
-// UPDATING APPROVAL STATUS to UNAPPROVED
-if (isset($_GET['unapprove'])) {
-    $the_comment_id = escape($_GET['unapprove']);
-
-    $query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = {$the_comment_id} ";
-
-    $unapprove_comment_query = mysqli_query($connection, $query);
-    header("Location: comments.php");
-}
-
-// UPDATING APPROVAL STATUS to APPROVED
-if (isset($_GET['approve'])) {
-    $the_comment_id = escape($_GET['approve']);
-
-    $query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = {$the_comment_id} ";
-
-    $approve_comment_query = mysqli_query($connection, $query);
-    header("Location: comments.php");
-}
-
 ?>
